PACEXI5-2249, PACEXI5-2251, DELIA-21546: Brought the Missing changes.
Reason for change: Noticed that https://gerrit.teamccp.com/#/c/94457/16 is not properly cherry-picked to stable2. There are 3 lines of change missed when manually cherry picked and that is leading to the crash. Fixed the issue by bringing in the left over changes.
Test Procedure: As per the linked tickets
Risks: Medium
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: I35c59536f9923f10e99019d3af2025e08a361131