PACEXI5-874, PACEXI5-1818 : The LNF Status was wronly updated; Fixed it now
Reason for change: Noticed that the loop is keep tring to connect to LNF/PRIVATE SSIDs... but whenever connection success, it didnt check whether the connection was to LNF or PRIVATE, it always updates the LNF Status as CONNECTED_PRIVATE. This is the root cause. Fixed it now by adding the conditional check.
Test Procedure: As per PACEXI5-1818
Risks: Low
Signed-off-by: Karunakaran A <Karunakaran_Amirthalingam@cable.comcast.com>

Change-Id: Id9615c5fffdda8596904078a064b1b1392c3ee74